Moving GTK Widgets to shoes/native/gtk/ directory
* Shoes3/shoes3#333
