+++ shoes/native/gtk.h
@@ -1,3 +1,4 @@
 #define GTK_CHILD(child, ptr) \
   GList *children = gtk_container_get_children(GTK_CONTAINER(ptr)); \
   child = children->data
+
