+++ shoes/native/gtk/gtkcomboboxtextalt.c
