+++ shoes/native/gtk/gtkcomboboxtextalt.h
