+++ shoes/native/gtk/gtkeditbox.h
