+++ shoes/native/gtk/gtkentryalt.c
