+++ shoes/native/gtk/gtkentryalt.h
