+++ shoes/native/gtk/gtkfixedalt.c
