+++ shoes/native/gtk/gtkprogressbaralt.c
