+++ shoes/native/gtk/gtkprogressbaralt.h
