+++ shoes/native/gtk/gtkspinner.h
